Consolidate PR skills into VCS-aware create-pr
